Enum tweetr::Outcome

pub enum Outcome {
    NoError,
    OverrideNoForce(String),
    RequiredFileFromSubsystemNonexistant {
        subsys: &'static str,
        fname: String,
    },
    RequiredDataFromSubsystemNonexistant {
        subsys: &'static str,
        desc: String,
    },
    TwitterAPIError(String),
    FileParsingFailed {
        desc: &'static str,
        errors: Vec<String>,
    },
}

Enum representing all possible values the application can fail.

Variants
NoError

No errors occured, everything executed correctly.

OverrideNoForce(String)

The specified file would need to be overriden but was not allowed to.

RequiredFileFromSubsystemNonexistant

The specified subsystem needs to be run beforehand to produce the specified file.

Fields

subsys: &'static str

The subsystem that needs to be run.

fname: String

The file the specified subsystem produces.

RequiredDataFromSubsystemNonexistant

The specified subsystem needs to be run beforehand to produce the specified data.

Fields

subsys: &'static str

The subsystem that needs to be run.

desc: String

The description of what needs to be produced.

TwitterAPIError(String)

The Twitter API returned an error.

FileParsingFailed

Failed to parse the specified file because of the specified errors.

Fields

desc: &'static str

The file that failed to parse.

errors: Vec<String>

The parsing errors that occured.

Methods

impl Outcome

fn print_error<W: Write>(&self, err_out: &mut W)

Get the executable exit value from an Outcome instance.

Examples

let mut out = Vec::new();
Outcome::OverrideNoForce("doctest".to_string()).print_error(&mut out);
assert_eq!(String::from_iter(out.iter().map(|&i| i as char)),
           "File \"doctest\" was not overriden to prevent data loss.\n\
            Pass --force to override it.\n".to_string());

fn exit_value(&self) -> i32

Get the executable exit value from an Outcome instance.

Examples

exit(Outcome::NoError.exit_value());