Module tweetr::ops::start_daemon

This module contains the functions used only by the start-daemon subsystem.

The flow of the start-daemon subsystem is as follows:

Initialisation:

Options::parse()
|> ops::start_daemon::verify()
|> ops::AppTokens::read()

Then, in a loop:

init_data
|> ops::User::read()
|> ops::QueuedTweet::read()
|> ops::start_daemon::tweet_indices_to_post()
|> ops::start_daemon::find_user_index_for_tweet()
|> ops::start_daemon::post_tweet()

Functions

find_user_index_for_tweet
Try to get the index of the user to post the given tweet.

post_tweet
Post the specified tweet on behalf of the specified user and application, optionally printing progress.

tweet_indices_to_post
Get the indices of tweets to post now from the provided batch based on whether thy've been posted already and the current time.

verify
Verify if, given the current configuration, it's permitted to continue with the subsequent steps of the start-daemon subsystem.